% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun.R
\name{align_bowtie2}
\alias{align_bowtie2}
\title{Aligns metagenomic shotgun sample against a reference}
\usage{
align_bowtie2(reads, index_basename, threads = 1,
  alignment_folder = "alignments", bam = TRUE, bowtie2_path = NULL,
  samtools_path = NULL)
}
\arguments{
\item{reads}{A data frame containing the paths to the forward and reverse
reads in fastq format. Needs to have at least columns "id" and "forward".
If the samples are paired you also need a column "reverse".}

\item{index_basename}{Path and basename of the index against which to align.}

\item{threads}{How many threads to use for bowtie2.}

\item{alignment_folder}{A folder to which to save log output and the
generated alignment.}

\item{bam}{Whether to output alignments in BAM format (requires samtools).}

\item{bowtie2_path}{Path to the bowtie executables.}

\item{samtools_path}{Path to samtools.}
}
\value{
A data frame mapping the alignments to the sample IDs. It will
 contain the following columns:
 \itemize{
 \item{id}{the id of the sample}
 \item{success}{whether alignment executed without error}
 \item{log}{the log output for the alignment}
 \item{alignment}{the file containing the alignment}
 }
}
\description{
This method uses bowtie2 and should not be too sensitive to the used pre-
processing.
}
\examples{
 NULL

}
