% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore.R
\name{align_long_reads}
\alias{align_long_reads}
\title{Align long reads (for instance from nanopore sequencing) to a reference
database.}
\usage{
align_long_reads(object, config)
}
\arguments{
\item{object}{An artifact or list of files.}

\item{config}{A configuration as obtained by \code{\link{config_align_short}}.}
}
\value{
A list with the generated alignments and some general diagnostics.
}
\description{
Align long reads (for instance from nanopore sequencing) to a reference
database.
}
