% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun.R
\name{align_short_reads}
\alias{align_short_reads}
\title{Aligns metagenomic shotgun reads against a reference.}
\usage{
align_short_reads(object, ...)
}
\arguments{
\item{object}{An artifact or list of files.}

\item{...}{Configuration arguments or a configuration object as obtained by
\code{\link{config_align}}.}
}
\value{
A list with the generated alignments and some general diagnostics.
}
\description{
This method uses bowtie2 and should not be too sensitive to the used pre-
processing.
}
\examples{
 NULL

}
