% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{annotate_contigs}
\alias{annotate_contigs}
\title{Annotate diamond database hits with transcript info.}
\usage{
annotate_contigs(matches, reference)
}
\arguments{
\item{matches}{Path to diamond output (*.m8).}

\item{reference}{Path to the reference sequences (*.fa.gz).}
}
\value{
The annotated hits as data table.
}
\description{
Annotate diamond database hits with transcript info.
}
