% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{annotate_silva}
\alias{annotate_silva}
\title{Annotate the sequences in the SILVA database file.}
\usage{
annotate_silva(reference)
}
\arguments{
\item{reference}{The reference fasta.}
}
\value{
The id and taxonomy of the sequences as a data table.
}
\description{
Annotate the sequences in the SILVA database file.
}
