% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.princomp.R
\name{anova.princomp}
\alias{anova.princomp}
\title{Component variation in PCA analysis}
\usage{
\method{anova}{princomp}(object, ...)
}
\arguments{
\item{object}{object of class \code{princomp}}

\item{...}{other arguments currently not supported}
}
\value{
A vector of length equal to a number of components existing in
\code{object}. Each element containes the percent of variance explained by
the corresponding component.
}
\description{
Calculate percent of explained variance for Principal Component Analysis
objects.
}
\details{
The function calculates percentages of variance explained for each component
resulting from Principal Component Analysis. The variance is equal to:
\deqn{100\frac{\sigma_i^2}{\sum_i \sigma_i^2}}{sigma^2 / sum( sigma^2 ) *
100} where \eqn{\sigma}{sigma} is equal to component's standard deviation.
}
\examples{
pcmodel <- princomp( USArrests, cor=TRUE )
pcmodel
anova(pcmodel) # variance prc.
}
\seealso{
\code{\link[stats]{princomp}}, \code{\link[stats]{anova}}
}
