% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{apply_types}
\alias{apply_types}
\title{Applies the specified types to a data frame-like object.}
\usage{
apply_types(df, types)
}
\arguments{
\item{df}{A data frame, data table or tibble.}

\item{types}{A data frame with two columns: name and type.}
}
\value{
The same frame with updated column types.
}
\description{
Applies the specified types to a data frame-like object.
}
\examples{
 NULL

}
