% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{as.matrix.mbquant}
\alias{as.matrix.mbquant}
\title{Convert a mbquant data table to a matrix.}
\usage{
\method{as.matrix}{mbquant}(x, ...)
}
\arguments{
\item{x}{the mbquant data table.}

\item{...}{additional arguments.}
}
\value{
A matrix with samples on the rows and taxa on the columns.
}
\description{
Convert a mbquant data table to a matrix.
}
\examples{
 NULL

}
