% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dada2.R
\name{as_phyloseq}
\alias{as_phyloseq}
\title{Converts the denoise artifact to a phyloseq object.}
\usage{
as_phyloseq(object, metadata = NULL)
}
\arguments{
\item{object}{The artifact returned by \code{\link{denoise}}.}

\item{metadata}{Sample metadata to add to the phyloseq object. Must have
rownames that match the sample names.}
}
\value{
A phyloseq object with additional annotations.
}
\description{
Converts the denoise artifact to a phyloseq object.
}
