% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.r
\name{assign_barcodes}
\alias{assign_barcodes}
\title{Assigns sample IDs from barcodes}
\usage{
assign_barcodes(index, barcodes)
}
\arguments{
\item{index}{Path to the FASTQ index file.}

\item{A}{vector of barcodes, one for each sample.}
}
\value{
A vector with an ID for each sample in the index file.
}
\description{
Assigns sample IDs from barcodes
}
\examples{
 NULL

}

