% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{association}
\alias{association}
\title{Run differential association tests between taxa counts and exogenous
factors.}
\usage{
association(ps, ...)
}
\arguments{
\item{ps}{A phyloseq object containing the taxa counts.}

\item{...}{A configuration as returned by \code{\link{config_association}}.}
}
\value{
An artifact containing the results.
}
\description{
Run differential association tests between taxa counts and exogenous
factors.
}
\examples{
 NULL

}
