% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{bin_coverage}
\alias{bin_coverage}
\title{Reads the coverage from a set of alignments and calculates binned
averages.}
\usage{
bin_coverage(object, ...)
}
\arguments{
\item{object}{An experiment data table as returned by any alignment method
like \code{\link{align_short_reads}} or \code{\link{align_long_reads}}.}

\item{...}{A configuration as generated by \code{\link{config_coverage}}.}
}
\value{
A list with the
}
\description{
Reads the coverage from a set of alignments and calculates binned
averages.
}
\examples{
 NULL
}
