% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bracken_to_phyloseq}
\alias{bracken_to_phyloseq}
\title{Converts BRACKEN results to a phyloseq object.}
\usage{
bracken_to_phyloseq(bracken, metadata = NULL, id_col = "sample_id")
}
\arguments{
\item{bracken}{A data.table containing the BRACKEN quantifications.}

\item{metadata}{A metadata in data.table format to use.}

\item{id_col}{Column in the metdata corresponding to the sample IDs
in the BRACKEN counts.}
}
\value{
A phyloseq object for the data.
}
\description{
Converts BRACKEN results to a phyloseq object.
}
\examples{
 NULL

}
