% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{build_lineage}
\alias{build_lineage}
\title{Build the lineage from a SILVA path string.
importFrom stringr str_split_fixed}
\usage{
build_lineage(silva_path, taxid, taxonomy)
}
\description{
Build the lineage from a SILVA path string.
importFrom stringr str_split_fixed
}
