% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore.R
\name{config_align_long}
\alias{config_align_long}
\title{Build a configuration for the long read alignment workflow.}
\usage{
config_align_long(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the long read alignment
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_align_long(index = "refs/mouse")
}
