% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun.R
\name{config_align_short}
\alias{config_align_short}
\title{Build a configuration for the short read alignment workflow.}
\usage{
config_align_short(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the short read alignment
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_count(index = "refs/mouse")
}
