% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{config_count}
\alias{config_count}
\title{Build a configuration for the transcript counting workflow.}
\usage{
config_count(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the transcript counting
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_count(reference = "refs/mouse.fna.gz")
}
