% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{config_coverage}
\alias{config_coverage}
\title{Build a configuration for the coverage workflow.}
\usage{
config_coverage(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the long read alignment
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_coverage(bin_width = 10000)
}
