% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dada2.R
\name{config_dada2}
\alias{config_dada2}
\title{Build a configuration for the DADA2 workflow.}
\usage{
config_dada2(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the DADA2 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_dada2(truncLen = c(240, 250))
}
