% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{config_demultiplex}
\alias{config_demultiplex}
\title{Build a configuration for the demultiplexing workflow.}
\usage{
config_demultiplex(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the demultiplexing workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_demultiplex(n = 1e4)
}
