% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{config_power}
\alias{config_power}
\title{Build a configuration for power analysis.}
\usage{
config_power(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in power analysis.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_power(fraction_differential = 0.1)
}
