% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{config_ptr}
\alias{config_ptr}
\title{Build a configuration for the PTR workflow.}
\usage{
config_ptr(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the long read alignment
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_ptr(span=0.5)
}
