% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ref.R
\name{config_reference}
\alias{config_reference}
\title{Build a configuration for the reference removal workflow.}
\usage{
config_reference(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the DADA2 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_reference(reference = "refs/mouse.fna.gz")
}
