% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication.R
\name{config_rep}
\alias{config_rep}
\title{Build a configuration for the rate workflow.}
\usage{
config_rep(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the long read alignment
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_rep(min_coverage = 0.8)
}
