% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{config_slimm}
\alias{config_slimm}
\title{Build a configuration for the SLIMM workflow.}
\usage{
config_slimm(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the long read alignment
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_slimm(bin_width = 2000)
}
