% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra.R
\name{config_sra}
\alias{config_sra}
\title{Build a configuration for the SRA submission workflow.}
\usage{
config_sra(...)
}
\arguments{
\item{...}{Any arguments are used to update the default configuration. See
the example below. Optional.}
}
\value{
A list with the parameters used in the transcript counting
 workflow.
}
\description{
This can be saved and passed on to others to ensure reproducibility.
}
\examples{
 config <- config_sra(date_col = "collection_date")
}
