% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore.R
\name{count_nanopore}
\alias{count_nanopore}
\title{Count nanopore hits to a 16S reference database.}
\usage{
count_nanopore(alignment_files)
}
\arguments{
\item{alignment_files}{Paths to BAM files.}
}
\value{
A data.table with sequence names, counts and sample name.
}
\description{
Count nanopore hits to a 16S reference database.
}
