% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{count_transcripts}
\alias{count_transcripts}
\title{Count alignment hits to a reference database.}
\usage{
count_transcripts(object, config)
}
\arguments{
\item{config}{A configuration as generated by \code{\link{config_count}}.}

\item{alignment_files}{Paths to BAM files.}
}
\value{
A list containing the used alignments and the transcript counts in
 `counts`.
}
\description{
This will correct for effective transcript lengths as done by almost any
good tool those days. So the returned counts are not correlated with
transcript lengths. By default an expectation maximization algorithm is
used to resolve multiple mappings of one read to many transcripts which
pretty much always happens in metagenomics data sets. The optimized
likelihodd function is very similar to the one in kallisto
(https://doi.org/10.1038/nbt.3519).
}
