% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{demultiplex}
\alias{demultiplex}
\title{Splits FASTQ files into individual samples.}
\usage{
demultiplex(object, config)
}
\arguments{
\item{reads}{A character vector containing the read files in fastq format.}

\item{index}{The index file containing the demultiplexed barcodes for each
read file.}

\item{out}{A folder to which to save the split fastq files.}

\item{ref}{A character vector or DNAStringSet containing the reference
barcodes.}

\item{max_ed}{Maximum allowed edit distance between the sequenced and
reference barcode.}

\item{n}{Maximum number of records to read in each iteration.}
}
\value{
A list containing the split files and matching statistics.
}
\description{
This function will take each of the sequences barcodes and tries to find
the corresponding barcode from the reference set. After that it writes new
fastq files containing only the reads for a single barcode.
}
\examples{
 NULL

}
