% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dada2.R
\name{denoise}
\alias{denoise}
\title{Runs a full DADA2 workflow.}
\usage{
denoise(object, config)
}
\arguments{
\item{object}{An experiment data table as returned by
\code{\link{find_read_files}} or a worflow object.}

\item{config}{A configuration file as returned by \code{\link{config_dada2}}.}
}
\value{
A list containing the workflow results:
\describe{
  \item{feature_table}{Matrix of sequence variant abundances. Samples are
   and sequences are columns.}
  \item{taxonomy}{Matrix of taxonomy assignments. Rows are sequences and
   columns are taxonomy ranks.}
  \item{errors}{The error profiles estimated by DADA2. One for each run and
   read direction (forward/reverse).}
  \item{error_plots}{Plotted error profiles. One for each run and
   read direction (forward/reverse).}
  \item{passed_reads}{How many reads were kept in each step. Rows are
    samples and columns are workflow steps.}
  \item{classified}{The proportion of sequence variants that could be
       where a specific taxa rank could be classified.}
}
}
\description{
This performs the following steps: \enumerate{
 \item{Learning the error rate for each run separately}
 \item{Inferring sequence variants for each run and sample}
 \item{Merging of feature tables across runs}
 \item{Consensus chimera removal}
 \item{Taxa assignment with Naive Bayes}
 \item{Species assignment by exact alignment}
 \item{Diagnostic plots of the error rates}
 }
You will usually want to preprocess the read files first with
\code{\link{preprocess}}.
Depending on your config the sequences might be represented by
an MD5 hash. In this case the taxa table has an additional `sequence` column
containing the real sequences.
}
