% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dePolish.R
\name{dePolish}
\alias{dePolish}
\title{Rip Polish letters of the diacritics}
\usage{
dePolish(x, encoding = NULL)
}
\arguments{
\item{x}{character; vector to be processed}

\item{encoding}{character; encoding used in \code{x}

The function substitutes (with \code{\link[=sub]{sub()}}) case-sensitively all Polish
letters with simple ASCII counterparts.

The \code{encoding} argument defines the way in which Polish letters are
encoded in \code{x}. If \code{NULL} and the platform is Windows the function
uses CP1250 encoding. If not on Windows the function uses
\code{\link[=iconv]{iconv()}} to translate the letters to default encoding on current
locale (via argument \code{from=""}).}
}
\value{
A character vector after substitutions.
}
\description{
The function substitutes all Polish letters in the provided character
vector with ASCII counterparts.
}
\seealso{
\code{\link[=sub]{sub()}}, \code{\link[=iconv]{iconv()}}
}
