% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Digitize.R
\docType{class}
\name{Digitize-class}
\alias{Digitize}
\alias{Digitize-class}
\title{Digitizing images}
\description{
Digitizing images
}
\section{Fields}{

\describe{
\item{\code{image}}{}

\item{\code{xpts,ypts}}{}

\item{\code{xinterval,yinterval}}{}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{initialize object}

\item{\code{plot(path = FALSE, col = "red", labels = TRUE, pos = 1, ...)}}{plot points}

\item{\code{start(..., append = TRUE, ask = FALSE, col = "red", type = "p")}}{Starts adding points}
}}

