% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{discretize}
\alias{discretize}
\title{Discretize all continuous variables in a data frame.}
\usage{
discretize(df, groups = 3)
}
\arguments{
\item{df}{A data frame-like object.}

\item{groups}{The number of groups into which to separate the data.}
}
\value{
The same data fram with updated columns.
}
\description{
This function will attempt to balance the groups so they contain similar
numbers of elements.
}
\examples{
 NULL

}
