% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{download_bacterial_transcripts}
\alias{download_bacterial_transcripts}
\title{Download the entire ENSEMBL transcript DB for bacteria.}
\usage{
download_bacterial_transcripts(out = "transcripts", np = 8,
  remove_redundant = TRUE)
}
\arguments{
\item{out}{Where to store the downloaded transcript files.}

\item{np}{How many parallel download processes to use.}

\item{remove_redundant}{Whether to remove alternative assemblies for the
same strain.}
}
\value{
A data table containing the downloaded species and assemblies.
}
\description{
Download the entire ENSEMBL transcript DB for bacteria.
}
