% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{download_ensembl}
\alias{download_ensembl}
\title{Helper function to download ENSEMBL transcripts (cdna).}
\usage{
download_ensembl(out = "transcripts", what = "cdna", collection, name,
  assembly)
}
\description{
Helper function to download ENSEMBL transcripts (cdna).
}
