% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockrobiota.R
\name{download_ftpdir}
\alias{download_ftpdir}
\title{Downloads all files from a given FTP directory}
\usage{
download_ftpdir(url, outdir, quiet = FALSE)
}
\arguments{
\item{url}{URL to the directory.}

\item{outdir}{The output directory (will be created).}

\item{quiet}{Boolean. Whether the download should happen quietly.}
}
\value{
A character vector of the downloaded filenames.
}
\description{
This can be used to download all file from a FTP directory.
}
\examples{
 NULL

}

