% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_organism.R
\name{download_index}
\alias{download_index}
\title{Builds index files for human sequence identification}
\usage{
download_index(where = "./bowtie2", genome_file = NA)
}
\arguments{
\item{where}{Where to save index files}

\item{genome_file}{The human genome to be used. NA means download a recent version.}
}
\value{
Nothing.
}
\description{
This function has to be run only once to enable analysis.
}
\examples{
 NULL

}
