% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_download.R
\name{download_sra}
\alias{download_sra}
\title{Download data from SRA.}
\usage{
download_sra(
  runtable,
  path = "sra/",
  retries = 3,
  threads = getOption("mc.cores", 1)
)
}
\arguments{
\item{runtable}{Path to a runtable as selected in the SRA interface.}

\item{path}{The folder into which to download the FASTQ files.}

\item{retries}{How often to try downloading again after failure.}

\item{threads}{Maximum number of parallel file downloads.}
}
\value{
The list of files with indicated download success.
}
\description{
To use this function first obtain a RunInfo Table as described at
\url{https://www.ncbi.nlm.nih.gov/Traces/study/?go=help}.
}
