% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_count}
\alias{em_count}
\title{Count transcripts using an Expectation Maximization (EM) algorithm.}
\usage{
em_count(txreads, txlengths, weights, ntx, nr, maxit = 1000L,
  reltol = 0.01, abstol = 0.01)
}
\arguments{
\item{txreads}{A minimal matrix where each row corresponds to an
alignment. The first column denotes transcript indices in [0, ntx-1] and
the second column denotes read indices in [0, nr -1].}

\item{txlengths}{The sequence lengths for each transcript. Ideally those
should be the effective transcript lengths meaning the overall number
of possible alignment start positions in a transcript.}

\item{weights}{Weights for individual alignments.}

\item{ntx}{The total number of unique transcripts.}

\item{nr}{The total number of unique reads.}

\item{maxit}{Maximum number of EM iterations.}

\item{reltol}{The relative tolerance for convergence.}

\item{abstol}{The absolute tolerance for convergence.}
}
\value{
A list with the following components.
    \describe{
     \item{p}{The length-normalized and read scaled transcript counts}
     \item{iterations}{The number of used EM iterations}
     \item{num_ecs}{The number of equivalence classes}
     \item{change}{The last osbserved absolute change in transcript counts}
    }
}
\description{
Count transcripts using an Expectation Maximization (EM) algorithm.
}
