% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERR260132}
\alias{ERR260132}
\title{Precomputed lineage calling and coverage example.}
\format{A SLIMM artifact.}
\source{
\url{https://www.ncbi.nlm.nih.gov/sra/?term=ERR260132}
}
\usage{
ERR260132
}
\description{
Generated from aligning SRA sample ERR260132
 after aligning
to the SLIMM database of 15K genbank genomes.
}
\keyword{datasets}
