% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etas.R
\name{etas}
\alias{etas}
\alias{etas.default}
\alias{etas.anova}
\alias{etas.lm}
\title{Calculate Eta coefficients}
\usage{
etas(object, ...)

\method{etas}{default}(object, fac, ...)

\method{etas}{anova}(object, ...)

\method{etas}{lm}(object, ...)
}
\arguments{
\item{object}{the R object}

\item{...}{arguments passed to other methods}

\item{fac}{vector for conditioning variable}
}
\value{
Values of eta and partial eta coefficients.
}
\description{
This generic function calculates Eta coefficients which are also known as
"Correlation ratios" or (the squared value) as the "Proportion of
explained variance".
}
\details{
The Eta coefficient (more specifically its squared value) has a
interpretation in terms of the proportion of explained variance.

In the decision theory the interpretation is related to the identification
problem that involves two variables: \eqn{y} and \eqn{x}. The task is two
identify the values of \code{y}.

The value of the \eqn{\eta^2}{Eta^2} is the proportion by which the error of
predicting values of \eqn{y} is reduced by using the information contained
in \eqn{x}.

For numeric vectors the function requires additional argument: a
vector of the same length as the first.  The result is a value of the
\eqn{\eta^2}{Eta^2} assuming that we want to predict the values of
\code{object} with the values of \code{fac} using the so called ``Type I
regression of means''.

For two variables \eqn{y} and \eqn{x} the \eqn{\eta}{Eta} is given by the
formula:

\deqn{\eta^2 = ( D^2(y) - E[D^2(y|x)] ) / D^2(y)}{Eta^2 = ( D^2(y) - E[D^2(y|x)] ) / D^2(y)}

For objects of class \code{anova} the function calculates the Eta's
and Partial Eta Squares for all effects in the given model. In this setting
the eta squares for the given effect are equal to:

\deqn{\frac{SS_{effect}}{SS_{total}}}{SSeffect / SStotal}

where \eqn{SS} are apropriate Sums of Squares.  The ``Partial Eta Squares''
for the given effect are equal to:

\deqn{\frac{SS_{effect}}{SS_{effect}+SS_{resid}}}{SSeffect / (SSeffect+SSresid)}

For objects of class \code{lm} the function is applied on the
result of calling \code{anova}.
}
\examples{
x1 <- rnorm(50)
x2 <- rnorm(50)
y <- 5 + 2*x1 + rnorm(50,0,2) + 3*x2 + rnorm(50,0,.5)

# method for numeric
etas( y, rep(1:2, each=25) )

# method for 'lm' which calls 'anova'
m <- lm( y ~ x1 + x2 )
etas(m)
}
