% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter low presence taxa from a count matrix.}
\usage{
filter_counts(counts, abundance = 10, presence = 0.1)
}
\arguments{
\item{counts}{A count matrix with samples as rows and taxa as columns.}
}
\value{
Count matrix with some taxa removed.
}
\description{
Filter low presence taxa from a count matrix.
}
\examples{
 NULL

}
