% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{find_illumina}
\alias{find_illumina}
\title{Find Illumina read files in a given directory.}
\usage{
find_illumina(directory, dirs_are_runs = FALSE)
}
\arguments{
\item{directory}{The directory in which to look.}

\item{dirs_are_runs}{Whether subdirctories indicate different sequencing
runs.}
}
\value{
A data table that contains the samples and their annotations.
}
\description{
Find Illumina read files in a given directory.
}
