% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{find_taxa}
\alias{find_taxa}
\title{Checks whether taxa from one taxonomy table are contained in another table.}
\usage{
find_taxa(taxa1, taxa2, level = "Species")
}
\arguments{
\item{taxa1}{First taxonomy table.}

\item{taxa2}{Second taxonomy table.}

\item{level}{At which level to compare. Must be column name in both tables.}
}
\value{
A data frame with two columns. Has as many rows as unique values in level.
\describe{
\item{level}{The unique values found for the specified taxa level.}
\item{found}{A boolean indicating whether the taxa were found in the second table.}
}
}
\description{
This function takes two taxonomy tables, than looks for each sample of the first
table in the second one.
}
\examples{
 NULL

}
