% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_timesheet.R
\name{git_timesheet}
\alias{git_timesheet}
\title{Identify work streaks from Git history}
\usage{
git_timesheet(dir = ".", threshold = 75)
}
\arguments{
\item{dir}{path to Git repo}

\item{threshold}{number of minutes. Commits further apart than \code{threshold} belong to different streaks}
}
\value{
Tibble with columns:
\itemize{
\item \code{streak} - Streak ID
\item \code{from} - Date-time, start of the work streak
\item \code{to} - Date-time, end of the work streak
}
}
\description{
Identify work streaks from Git history
}
\details{
For every commit the function estimates the "commit time" which is
the amount of time spent working towards that commit. Current approach is rather naive:
\enumerate{
\item Break the data into streaks using \code{threshold}
\item For every streak:
\itemize{
\item All but first commit get commit time equal to the time difference from the previous commit.
\item For the first commit the commit time is the mean of all commits in the streak.
}
\item For streaks of length one the commit time is the mean of all commit times for the repo.
}
}
