% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github}
\alias{github}
\alias{github_details}
\alias{session_info}
\title{Various GitHub tools}
\usage{
github_details(x, summary_text = "Summary")

session_info()
}
\arguments{
\item{x}{character; text to be wrapped in Details}

\item{summary_text}{character; the title of the Details}
}
\description{
\itemize{
\item \code{github_details()} - Wrap input with markup proper for
the "Details" element of a GitHub issue etc.
}

\itemize{
\item \code{session_info()} - Run \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}},
format the output and copy to clipboard for easy pasting on GitHub.
}
}
\examples{
ch <- c("a", "b")
cat(github_details(ch))
}
