% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.symmetric.R
\name{is.symmetric}
\alias{is.symmetric}
\title{Test for a symmetric matrix}
\usage{
is.symmetric(m)
}
\arguments{
\item{m}{a matrix to be tested}
}
\value{
Logical, whether the matrix is symmetric or not.
}
\description{
Test whether the supplied matrix is symmetric around diagonal
}
\details{
I the matrix \code{m} is not a square matrix the function returns
\code{FALSE} and issues a warning.

The testing boils down to comparing lower and upper triangles of the matrix
\code{m} using the function \code{identical}.
}
\examples{
# create some matrix
m <- matrix( c(0, 1, 1, 0), ncol=2 )
m
is.symmetric(m) # TRUE

m1 <- m
m1[2,1] <- 0
m1
is.symmetric(m1) # FALSE

# supplying a non-square matrix
m2 <- matrix( 1:6, ncol=2 )
is.symmetric(m2) # FALSE with warning

# clean-up
rm(m, m1, m2)

}
\seealso{
See \code{\link{identical}}, \code{\link{matrix}},
\code{\link{upper.tri}}. Also \code{\link[sna]{symmetrize}} in package
\pkg{sna}.
}

