% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isGraphical.R
\name{isGraphical}
\alias{isGraphical}
\title{Test if a degree sequence isGraphical}
\usage{
isGraphical(x, method = c("tv", "eg"))
}
\arguments{
\item{x}{integer vector, degree sequence to be tested}

\item{method}{character, type of algorithm to be used, defaults to "tv", see
Details}
}
\value{
Logical, whether the sequence is graphical or not.
}
\description{
Given a vector of integer values check whether it is \emph{graphical}, i.e.
whether there exists an undirected graph with a corresponding degree
sequence.
}
\details{
A sequence of non-negative integers \eqn{a_1, a_2, ..., a_p}{a[1], a[2],
..., a[p]}. is \emph{graphical} if there exists a graph with the
corresponding degree sequence.

This function takes such a sequence as an argument \code{x} and performes
the check.

Two algorithms are implemented. The first one due too Erdos and Gallai
(1960), the second, a bit more efficient and the default, is due too
Tripathi and Vijay (2003). The method can be selected with the \code{method}
argument.
}
\examples{
### Test both methods on some simple graphs
testit <- function(d)
{
    rval <- c( eg = isGraphical(d, "eg"),
        tv = isGraphical(d, "tv") )
    if(!all(rval))
    {
        print(rval)
        stop("failed")
    }
    rval
}
# graphs of size 3
testit( c(0,0,0) )
testit( c(0,1,1) )
testit( c(1,2,1) )
testit( c(2,2,2) )

}
\references{
P. Erdos, T. Gallai (1960) Graphs with prescribed degree of vertices
(Hungarian), Mat. Lapok 11 264--274.

A. Tripathi and Sujith Vijay (2003) A note on a theorem of Erdos and Gallai,
Discrete Mathematics 265 417--420
}

