% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{layout}
\alias{layout}
\title{Generate a plate layout from a sample manifest.}
\usage{
layout(manifest, ...)
}
\arguments{
\item{manifest}{A data frame-like object containing the samples.}

\item{...}{other parameters passed to \code{\link{config_layout}}.}
}
\value{
An artifact with the annotated manifest as well as a plate map.
}
\description{
Generate a plate layout from a sample manifest.
}
