% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Locator.R
\docType{class}
\name{Locator-class}
\alias{Locator}
\alias{Locator-class}
\title{Reference-class implementation of locator()}
\description{
Reference-class implementation of locator()
}
\section{Fields}{

\describe{
\item{\code{pts}}{a list of components 'x' and 'y' containg the coordinates of the points}
}}
\section{Methods}{

\describe{
\item{\code{identify()}}{identify points}

\item{\code{initialize()}}{initialize object}

\item{\code{plot(path = FALSE, col = "red", labels = TRUE, pos = 1, ...)}}{plot points}

\item{\code{remove(ind)}}{removing points}

\item{\code{reorder(...)}}{reorder points}

\item{\code{start(..., append = TRUE, ask = FALSE, col = "red", type = "p")}}{Starts adding points}
}}

