% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lss.R
\name{lss}
\alias{lss}
\title{List object names along with classes and sizes}
\usage{
lss(graph = FALSE, env = .GlobalEnv, ...)
}
\arguments{
\item{graph}{logical, should a dotchart of sizes by produced}

\item{env}{environment to scan for objects, defaults to Global Environment}

\item{...}{other arguments passed to \code{ls}}
}
\value{
A data frame with two columns \code{class} and \code{size} containing the
class and size (in bytes) of the objects found in the environment
\code{env}.
}
\description{
List, and optionally plot, all the objects in the specified environment
along with their class attribute and size in bytes.
}
\details{
The function scans the environment supplied by \code{env} with the \code{ls}
function for objects. Then the classes of these objects are retrieved with
\code{object.size}.

If \code{graph} is \code{TRUE} then additionally a dotchart with showing
object sizes is produced.
}
\examples{

# make some environment
e <- new.env()
# create some objects in it
assign("x", rnorm(10), envir=e)
assign("d", data.frame(a=1:10, b=10:1), envir=e)
assign("l", letters, envir=e)

# table of contents
lss(env=e)

# with a chart
lss(TRUE, env=e)

# clean-up
rm(e)

}
\seealso{
\code{ls} for listing the objects, \code{object.size} for
calculating the memory size of an object, \code{dotchart} for making
dotcharts
}
\keyword{environment}
