\name{mbtools-package}
\alias{mbtools-package}
\alias{mbtools}
\docType{package}
\title{Collection of development tools for R}
\description{
This package is a collection of utility function for working with \R
and developing for \R.
}
\author{
Michal Bojanowski\cr
\email{michal2992@gmail.com}\cr
\url{http://www.bojanorama.pl}
}
\keyword{package}
