% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtools-package.R
\docType{package}
\name{mbtools}
\alias{mbtools}
\alias{mbtools-package}
\title{mbtools.}
\description{
`mbtools` includes several helper functions for analysis and benchmarking
of microbial community analysis.
}
