% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{merge_transcripts}
\alias{merge_transcripts}
\title{Merge transcripts into a single database}
\usage{
merge_transcripts(
  transcript_files,
  transcript_folder,
  what = "cdna",
  out = "ensembl_transcripts.fa.gz"
)
}
\arguments{
\item{transcript_files}{A data table as returned by
\code{\link{download_bacterial_transcripts}}.}

\item{transcript_folder}{In which folder to look for transcripts.}

\item{what}{The type of database, either "cdna" or "protein".}

\item{out}{The filename for the DB. Will be compressed and saved in the
transcript folder.}
}
\value{
The transcript counts for each file.
}
\description{
Merge transcripts into a single database
}
