% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_deps.R
\name{missing_deps}
\alias{missing_deps}
\title{Check whether there are missing dependencies}
\usage{
missing_deps(deps = c("Depends", "Imports", "LinkingTo"), ...)
}
\arguments{
\item{deps}{character vector of elements from "Depends", "Suggests",
"Imports", "Enhances", or "LinkingTo"}

\item{...}{arguments passed to \code{\link[cranet:pkgnet]{cranet::pkgnet()}}}
}
\value{
Character vector of names of missing packages with and igraph object
as a \code{"graph"} attribute.
}
\description{
Check whether there are missing dependencies
}
\details{
First \code{\link[=installed.packages]{installed.packages()}} is called to get a matrix of all the
installed packages. One can use \code{...} to pass e.g. \code{lib.loc} to
\code{\link[=installed.packages]{installed.packages()}} to restrict attention to selected library tree.
Obtained output is passed to \code{\link[cranet:pkgnet]{cranet::pkgnet()}} to build the dependency
graph. Use \code{deps} argument to control what kind of dependency relations are
to be used. By default only "hard" dependencies are looked at. Finally the
function compares the packages in the graph to the list of installed
packages. Any packages present in the first but absent in the second should
be installed but are not. The vector of names of these is returned.
}
\seealso{
\code{\link[cranet:pkgnet]{cranet::pkgnet()}}
}
