% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_plot}
\alias{mock_plot}
\title{Creates a plot of measured taxa quantifications vs. reference quantification.}
\usage{
mock_plot(taxa1, taxa2)
}
\arguments{
\item{taxa1}{Measured taxonomy quantities.}

\item{taxa2}{Reference taxonomy quantities.}
}
\value{
A ggplot2 plot.
}
\description{
Creates a plot of measured taxa quantifications vs. reference quantification.
}
\examples{
 NULL

}

