% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{mothur_to_dada}
\alias{mothur_to_dada}
\title{Converts taxa annotations from mothur format to dada2 format.}
\usage{
mothur_to_dada(seq_file, taxa_file, out = "taxonomy.fa.gz")
}
\arguments{
\item{seq_file}{A fasta file containing the (cluster) sequences.}

\item{taxa_file}{A tab-separated file with IDs on the first column and
taxonomy in the second column.}

\item{out}{Filename for the compressed output file.}
}
\value{
Nothing.
}
\description{
Converts taxa annotations from mothur format to dada2 format.
}
\examples{
 NULL

}
