% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{normalize}
\alias{normalize}
\title{Normalize a set of read counts across samples}
\usage{
normalize(counts, method = "poscounts")
}
\arguments{
\item{counts}{A count matrix with samples as rows or mbquant object.}

\item{method}{The method to use. Defaults to "poscounts".}
}
\value{
An object containing normlized read counts.
}
\description{
Uses DESeq2's size factor estimation.
}
\examples{
 NULL

}
