% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullModel.R
\name{nullModel}
\alias{nullModel}
\alias{print.nullModel}
\title{Fit two-level null random intercept model}
\usage{
nullModel(y, g, data = NULL)

\method{print}{nullModel}(x, ...)
}
\arguments{
\item{y}{numeric, dependent variable}

\item{g}{numeric, grouping factor}

\item{data}{data frame, optional dataset where the variables should be
looked for}

\item{x}{object of class "nullModel"}

\item{...}{other arguments passed to/from other methods}
}
\value{
The function returns a list of S3 class "nullModel" which
contains the following components:
\describe{
\item{yname}{character, name of the dependent variable}
\item{gname}{character, name of the grouping factor}
\item{n}{numeric, number of groups}
\item{m}{numeric, total number of observations}
\item{wgvar}{numeric, pooled within-group variance}
\item{bgvar}{numeric, between-group variance}
\item{tau}{numeric, estimate of the population between-group variance}
\item{intcor}{numeric, intra-class correlation coefficient}
}
}
\description{
Fit a null two-level random intercept model using closed-form formulas
provided by Snijders nd Bosker (1999)
}
\details{
The function fits a random effect model of the form:

\deqn{Y_{ij} = \mu + U_j + R_{ij}}{Y_ij = mu + U_j + R_ij}

where \eqn{Y_{ij}}{Y_ij} is the value of the dependent variable (\code{y})
for i-th object within j-th group, \eqn{\mu}{mu} is the population mean,
\eqn{U_j} is the effect specific to group j, and \eqn{R_{ij}}{R_ij} is the
within-grup error.

The function estimates between-group and pooled within-group variances as
well as the intra-class correlation coefficient.  The closed-form
expressions are provided by Snijders and Bosker (1999, p. 18--21).
}
\examples{

\dontrun{
# fit null model with lmer
library(lme4)
system.time(mm <- lmer( Reaction ~ 1 | Subject, sleepstudy ))
# fit with closed-form formulas
system.time(m <- nullModel( sleepstudy$Reaction, sleepstudy$Subject ))
mm
m
}

}
\references{
Snijders, T.A.B., Bosker, R. (1999) "Multilevel Analysis. An
Introduction to Basic and Advanced Multilevel Modelling". London: Sage
}
\seealso{
\code{lmer} in package \pkg{lme4} for general fitting of mixed-effects
models.
}
