% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{orlitsky}
\alias{orlitsky}
\title{Orlitsky's diminishing attenuation estimator (q2/3).}
\usage{
orlitsky(counts)
}
\arguments{
\item{counts}{A vector of counts for which to approximate discrete
probablities.}
}
\value{
A named vector `p` assigning a probability to each event in data.
 Those do not sum up to one since there is also a remaining probability to
 observe a new event given as p(new) = 1 - sum(p).
}
\description{
Orlitsky's diminishing attenuation estimator (q2/3).
}
\examples{
 x <- sample(1:10, 100, replace=TRUE)
 p <- orlitsky(x)
}
