% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{parse_ensembl_id}
\alias{parse_ensembl_id}
\title{Parse annotations from an ENSEMBL id}
\usage{
parse_ensembl_id(id)
}
\arguments{
\item{id}{The id to be parsed.}

\item{A}{data table containing the transcript id with annotations.}
}
\description{
Parse annotations from an ENSEMBL id
}
