% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{peak_to_through}
\alias{peak_to_through}
\title{Estimate replication rates by the peak-to-through ratio.}
\usage{
peak_to_through(object, ...)
}
\arguments{
\item{object}{An artifact containing a coverage map as returned by
\code{link{slimm}}.}

\item{...}{Additional configuration parameters or a config object as
returned by \code{\link{config_ptr}}.}
}
\value{
An artifact containing the peak-to-through ratios as well as the
 smoothed coverage profiles.
}
\description{
Implements the method from Korem et. al. (DOI: 10.1126/science.aac4812).
}
