% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{plot_bb_fits}
\alias{plot_bb_fits}
\title{Plot the Betabinomial fits for all taxa in a phyloseq object}
\usage{
plot_bb_fits(ps, bins = NULL)
}
\arguments{
\item{ps}{A phyloseq object containing data for a reference experiment.
This should assume absence of any differential effect (all variation is
random).}

\item{bins}{number of bins to use for histograms. Defaults to Sturges rule.}
}
\value{
A ggplot2 plot object shwoing the fits for each taxon.
}
\description{
Plot the Betabinomial fits for all taxa in a phyloseq object
}
