% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{plot_entropy}
\alias{plot_entropy}
\title{Plots the distribution of base entropy for each cycle.}
\usage{
plot_entropy(qp)
}
\arguments{
\item{cycles}{A quality profile as returned by
\code{\link{quality_profile}}.}
}
\value{
A ggplot2 plot mapping the read positions to mean quality for each
 sample. Will be facetted into forward and reverse if applicable.
}
\description{
Plots the distribution of base entropy for each cycle.
}
