% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockrobiota.R
\name{plot_found}
\alias{plot_found}
\title{Creates a barplot of found taxa on each level of the taxonomy.}
\usage{
plot_found(taxa1, taxa2)
}
\arguments{
\item{taxa1}{First taxonomy table.}

\item{taxa2}{Second taxonomy table.}
}
\value{
A bar plot denoting the percentage of correctly identified taxa.
}
\description{
Creates a barplot of found taxa on each level of the taxonomy.
}
\examples{
 NULL

}

