% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{plot_lengths}
\alias{plot_lengths}
\title{Plots the distribution of sequence lengths with acceptable quality.}
\usage{
plot_lengths(qp, min_score = 10)
}
\arguments{
\item{min_score}{Smallest acceptable score. Defaults to 10 (10% per base
error).}

\item{cycles}{A quality profile as returned by
\code{\link{quality_profile}}.}
}
\value{
A ggplot2 plot mapping the read positions to mean quality for each
 sample. Will be facetted into fowrward and reverse if applicable.
}
\description{
Plots the distribution of sequence lengths with acceptable quality.
}
