% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{plot_qualities}
\alias{plot_qualities}
\title{Plots the quality profile for an entire experiment.}
\usage{
plot_qualities(qp)
}
\arguments{
\item{cycles}{A quality profile as returned by
\code{\link{quality_profile}}.}
}
\value{
A ggplot2 plot mapping the read positions to mean quality for each
 sample. Will be facetted into fowrward and reverse if applicable.
}
\description{
Plots the quality profile for an entire experiment.
}
