% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockrobiota.R
\name{plot_quants}
\alias{plot_quants}
\title{Creates a scatter plot of found versus real quantities.}
\usage{
plot_quants(taxa1, taxa2)
}
\arguments{
\item{taxa1}{First taxonomy table.}

\item{taxa2}{Second taxonomy table.}
}
\value{
A scatter plot having the measured quantities on the x-axis and
 true quantities on the y-axis.
}
\description{
Creates a scatter plot of found versus real quantities.
}
\examples{
 NULL

}

