% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_taxa}
\alias{plot_taxa}
\title{Plots relative distribution for taxa across samples.}
\usage{
plot_taxa(ps, level = "Phylum", sort = TRUE, max_taxa = 12,
  only_data = FALSE)
}
\arguments{
\item{ps}{A phyloseq data set.}

\item{level}{The taxonomy level to use. Defaults to phylum.}

\item{sort}{Whether to sort taxa by abundance across all samples.}

\item{max_taxa}{Maximum number of different taxa to plot. If more than 12
there is probably no color scale that can visualize them.}

\item{only_data}{Only get the raw data for the plot as a data table.}
}
\value{
Nothing or a data.table containing the relative abundances.
}
\description{
Plots relative distribution for taxa across samples.
}
\examples{
 NULL

}
