% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{power_analysis}
\alias{power_analysis}
\title{Run a power analysis for microbe abundances from a reference sample.}
\usage{
power_analysis(ps, ...)
}
\arguments{
\item{ps}{A phyloseq object containing data for a reference experiment.
This should assume absence of any differential effect (all variation is
random).}

\item{...}{A configuration as described by \code{\link{config_power}}.}
}
\value{
An artifact with an entry `power` that quantifies the power
 over many n and effect size.
}
\description{
Run a power analysis for microbe abundances from a reference sample.
}
