% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Runs preprocessing of sequencing reads.}
\usage{
preprocess(object, config)
}
\arguments{
\item{object}{An experiment data table as returned by
\code{\link{find_read_files}} or a worflow object.}

\item{config}{A configuration file as returned by
\code{\link{config_preprocess}}.}
}
\value{
A list containing the workflow results:
\describe{
  \item{passed_reads}{How many reads were kept in each step. Rows are
    samples and columns are workflow steps.}
  \item{files}{Preprocessed sequencing files list.}
}
}
\description{
Runs preprocessing of sequencing reads.
}
