% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{quality_profile}
\alias{quality_profile}
\title{Get the quality profiles and base calls for each cycle across all
input files.}
\usage{
quality_profile(files, n = 10000)
}
\arguments{
\item{files}{A data table containing the files as returned by
\code{\link{find_read_files}}.}

\item{n}{How many reads to sample from each file. If a file has less
all of them will be used.}
}
\value{
A list with two data tables giving the sampled metrics per cycle /
 base pair:
 \describe{
   \item{qualities}{The quality scores for each cycle and sample.}
   \item{bases}{The base calls for each cycle and sample.}
 }
}
\description{
Get the quality profiles and base calls for each cycle across all
input files.
}
