% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_describe.R
\name{rd_describe}
\alias{rd_describe}
\alias{rd_describe.default}
\alias{rd_describe.data.frame}
\title{RD generator: describe}
\usage{
rd_describe(object, ...)

\method{rd_describe}{default}(object, description, ...)

\method{rd_describe}{data.frame}(object, ...)
}
\arguments{
\item{object}{R object}

\item{...}{other arguments passed to/from other methods}

\item{description}{Character vector of term descriptions}
}
\value{
Nothing, but RD syntax is \code{\link[=cat]{cat()}}ed.
}
\description{
Generate RD markup for the definition list via the \verb{\\describe} command.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The default method expects character vector of
"terms" and a character vector \code{description} of the same length containing
the term descriptions.

\item \code{data.frame}: Generate a \verb{\\describe\{\}} markup with a list of
variable names together with variable labels defined (as in the
\strong{labelled} package).
}}

\concept{RD generators}
