% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_blast}
\alias{read_blast}
\title{Read a blast hit/alignment file.}
\usage{
read_blast(matches)
}
\arguments{
\item{matches}{The file containing the matches like it is returned by
blast or diamond.}
}
\value{
The hits as a data table.
}
\description{
Read a blast hit/alignment file.
}
