% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{read_length}
\alias{read_length}
\title{Heuristic to get the median read length from a BAM file.access}
\usage{
read_length(bam_file, n = 100)
}
\arguments{
\item{bam_file}{Path to a bam file.}

\item{n}{How many alignments to use.}
}
\value{
The median read length in the file.
}
\description{
Heuristic to get the median read length from a BAM file.access
}
