% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{read_slimm}
\alias{read_slimm}
\title{Reads SLIMM output to a data table.}
\usage{
read_slimm(profile)
}
\arguments{
\item{reports}{Where the slimm output is stored.}
}
\value{
A data table counting reads and relative abundances for all found
 bacteria for several taxonomic ranks.
}
\description{
Reads SLIMM output to a data table.
}
\examples{
 NULL

}
