% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{read_slimm_coverage}
\alias{read_slimm_coverage}
\title{Reads SLIMM coverage to a data table.}
\usage{
read_slimm_coverage(cofile, bin_width)
}
\arguments{
\item{cofile}{A SLIMM coverage file.}

\item{bin_width}{The width of the bin in base pairs.}
}
\value{
A data table containing the taxonomy and coverages.
}
\description{
Reads SLIMM coverage to a data table.
}
\examples{
 NULL
}
