% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yaml.R
\name{read_yaml}
\alias{read_yaml}
\alias{read_yaml_file}
\title{Parses YAML file with a simple R handler}
\usage{
read_yaml(string, ...)

read_yaml_file(input, ...)
}
\arguments{
\item{string}{YAML string to read}

\item{...}{other arguments}

\item{input}{YAML file to read}
}
\description{
Parses YAML file with a simple R handler
}
\examples{
y <- "num: 1
key0:
  key1: value1
key2:
  key22: value22
rcode_num: !r 1:5
rcode_ch: !r paste(LETTERS[1:10], collapse=', ')
date: !r Sys.Date()
"
read_yaml(y)
}
