% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{reads_to_phyloseq}
\alias{reads_to_phyloseq}
\title{Converts read count data to a phyloseq object.}
\usage{
reads_to_phyloseq(counts, feature_col, metadata = NULL, id_col = "sample_id")
}
\arguments{
\item{counts}{A data.table containing at least columns "reads",
"sample_id", and features.}

\item{feature_col}{Column in the count table containing the features.}

\item{metadata}{A metadata in data.table format to use.}

\item{id_col}{Column in the metadata corresponding to the sample IDs
in the read counts.}
}
\value{
A phyloseq object for the data.
}
\description{
Converts read count data to a phyloseq object.
}
\examples{
 NULL

}
