% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_human.r
\name{remove_human}
\alias{remove_human}
\title{Removes human sequences from a set of reads.}
\usage{
remove_human(reads, out, barcode_files = NA, index = "./bmtagger")
}
\arguments{
\item{reads}{A character vector containing the read files in fastq format.}

\item{out}{A folder to which to save the filtered fastq files.}

\item{barcode_files}{Additional barcode file that should be filtered as well.}

\item{index}{Where to find the previously generated index files.}
}
\value{
A numeric vector with two entries. The number of sequences after
 filtering (non-human), and the number of removed sequnces (human).
}
\description{
This function uses bmtagger to identify and remove humans sequences. It requires
prior construction of the indices via \link{build_index}.
}
\examples{
 NULL

}

