% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_organism.R
\name{remove_organism}
\alias{remove_organism}
\title{Removes sequence contamination from another target organism (for instance
human sequences).}
\usage{
remove_organism(reads, index = NA, out, organism = "hg19",
  where = "./bowtie2")
}
\arguments{
\item{reads}{A character vector containing the read files in fastq format.}

\item{index}{Additional barcode file that should be filtered as well.}

\item{out}{A folder to which to save the filtered fastq files.}

\item{organism}{The base name for the index. Defaults to human (hg19).}

\item{where}{Where to find the previously generated index files.}
}
\value{
A numeric vector with two entries. The number of sequences after
 filtering (non-human), and the number of removed sequences (human).
}
\description{
This function uses bowtie2 to align and identify hits.
This requires a pre-built index, for instance by using \link{download_index}.
}
\examples{
 NULL

}
