% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ref.R
\name{remove_reference}
\alias{remove_reference}
\title{Removes sequence that map to a given reference.}
\usage{
remove_reference(reads, index = NA, out, organism = "hg19",
  where = "./bowtie2", keep_bam = FALSE)
}
\arguments{
\item{reads}{A character vector containing the read files in fastq format.
Can be generated using \link{\code{find_illumina}}.}

\item{index}{Additional barcode file that should be filtered as well.}

\item{out}{A folder to which to save the filtered fastq files.}

\item{where}{Where to find the previously generated index files.}

\item{keep_bam}{Whether to keep the alignment. If not FALSE should be a
string indicating the path to the output bam file.}

\item{reference}{The base name for the index. Defaults to human (hg19).}
}
\value{
A numeric vector with two entries. The number of sequences after
 filtering (non-mapped), and the number of removed sequences (mapped).
}
\description{
This can be used for a variety of applications the most common ones are:
\itemize{
  \item removing sequences from the host
  \item removing ribosomal sequences
  \item removing contaminants
}
This function uses bowtie2 to align and identify hits.
This requires a pre-built index, for instance by using \link{download_index}.
}
\examples{
 NULL

}
