% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication.R
\name{replication_rates}
\alias{replication_rates}
\title{Estimate replication rates by the iRep method.}
\usage{
replication_rates(object, ...)
}
\arguments{
\item{object}{An artifact containing a coverage map as returned by
\code{link{slimm}}.}

\item{...}{Additional configuration parameters or a config object as
returned by \code{\link{config_rep}}.}
}
\value{
An artifact containing the peak-to-through ratios as well as the
 smoothed coverage profiles.
}
\description{
Implements the method from Brown et. al.
(https://dx.doi.org/10.1038%2Fnbt.3704).
}
