% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruler.R
\name{ruler}
\alias{ruler}
\title{Measuring number of columns of textual output}
\usage{
ruler(width = 80)
}
\arguments{
\item{width}{numeric length of the ruler}
}
\description{
Measuring number of columns of textual output
}
