% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{select}
\alias{select}
\title{Very simple select for mbtools artifacts. Just gets a particular entry.}
\usage{
select(object, entry)
}
\arguments{
\item{object}{A mbtools artifact returned from a workflow step.}

\item{entry}{The name of the entry to get.}
}
\value{
The requested entry.
}
\description{
Very simple select for mbtools artifacts. Just gets a particular entry.
}
