% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_info.R
\name{session_info}
\alias{session_info}
\title{Session info for GitHub}
\usage{
session_info()
}
\description{
Run \code{sessioninfo::session_info()}, format the output and copy to clipboard
for easy pasting on GitHub.
}
