% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimm.R
\docType{data}
\name{slimm_files}
\alias{slimm_files}
\title{The SLIMM taxonomy databases}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 3 columns.}
\usage{
slimm_files
}
\description{
The SLIMM taxonomy databases
}
\keyword{datasets}
