% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceall.R
\name{sourceall}
\alias{sourceall}
\title{Source all R-code files that match the pattern}
\usage{
sourceall(dir = getwd(), pattern = ".*\\\\.R", ...)
}
\arguments{
\item{dir}{character, the directory to look for the files. Defaults to the
working directory}

\item{pattern}{character, the regular expression that defines the pattern of
the files to be sourced. Defaults to all files with ".R" extension.}

\item{\dots}{other arguments passed to \code{source()}}
}
\description{
This function sources the files in a specified directory that follow the
supplied pattern.
}
\details{
All files in the \code{dir} directory are sourced in alphabetical order via
\code{\link{source}}.
}
\seealso{
\code{\link{source}}
}
\keyword{file}
\keyword{programming}
