% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{split_barcodes}
\alias{split_barcodes}
\title{Splits FASTQ files into individual samples.}
\usage{
split_barcodes(reads, index, out, ref, n = 1e+05, max_ed = 1)
}
\arguments{
\item{reads}{A character vector containing the read files in fastq format.}

\item{index}{The index file containing the demultiplexed barcodes for each
read file.}

\item{out}{A folder to which to save the split fastq files.}

\item{ref}{A character vector or DNAStringSet containing the reference
barcodes.}

\item{n}{Maximum number of records to read in each iteration.}

\item{max_ed}{Maximum allowed edit distance between the sequenced and
reference barcode.}
}
\value{
A numeric vector containing three entries, where the first defines the
 reads that are kept.
 \itemize{
 \item{The number of reads that could be mapped uniquely.}
 \item{The number of reads for which no match was found.}
 \item{The number of reads for which more than one reference match was found.}
 }
}
\description{
This function will take each of the sequences barcodes and tries to find
the corresponding barcode from the reference set. After that it writes new
fastq files containing only the reads for a single barcode.
}
\examples{
 NULL

}
