% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sra_files}
\alias{sra_files}
\title{List samples and read files in a directory.}
\usage{
sra_files(path)
}
\arguments{
\item{path}{Where to look for files.}
}
\value{
A data frame with columns "id", "forward" and optionally "reverse".
}
\description{
List samples and read files in a directory.
}
\examples{
 NULL

}
