% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra.R
\name{sra_submission}
\alias{sra_submission}
\title{Prepare submission files for the NCBI sequence read archive (SRA).}
\usage{
sra_submission(object, ...)
}
\arguments{
\item{object}{An experiment data table as returned by any alignment method
like \code{\link{align_short_reads}} or \code{\link{align_long_reads}} .}

\item{...}{Configuration arguments or a configuration as generated by
\code{\link{config_sra}}.}
}
\value{
A list containing the used alignments and the transcript counts in
 `counts`.
}
\description{
Prepare submission files for the NCBI sequence read archive (SRA).
}
