% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcenv.R
\name{srcenv}
\alias{srcenv}
\title{Return an environment with objects defined in a script file}
\usage{
srcenv(file, ...)
}
\arguments{
\item{file}{R file to source}

\item{...}{other arguments passed to \code{sys.source}}
}
\description{
Return an environment with objects defined in a script file
}
