% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{standardize}
\alias{standardize}
\title{Standardize all continuous columns of a data frame.}
\usage{
standardize(df)
}
\arguments{
\item{df}{A data frame-like object.}
}
\value{
The same data frame with standardized columns.
}
\description{
Standardize all continuous columns of a data frame.
}
\examples{
 NULL

}
