% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{taxa_metrics}
\alias{taxa_metrics}
\title{Calculates what percentage of taxa was found in a reference set.}
\usage{
taxa_metrics(tax_table, ref)
}
\arguments{
\item{tax_table}{Measured taxonomy table.}

\item{ref}{Reference taxonomy table.}
}
\value{
A data frame denoting performance metrics for taxa identification.
}
\description{
Note that both arguments must of class "taxonomyTable" from the
phyloseq package which can be obtained with the \code{tax_table}
method.
}
\examples{
 NULL

}
