% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{taxa_quants}
\alias{taxa_quants}
\title{Compares taxa quantification from a measurement to a reference ground truth.}
\usage{
taxa_quants(ps, ref, normalize = FALSE)
}
\arguments{
\item{ps}{A phyloseq object describing the measurements}

\item{ref}{A phyloseq object describing the reference set. Can be obtained
from \code{\link{mockrobiota}}.}

\item{normalize}{Whether to normalize taxa counts first.}
}
\value{
A data frame with the following columns:
 \describe{
 \item{level}{The taxonomy level for the entry.}
 \item{name}{The taxonomy.}
 \item{sample}{The sample name for the quantification.}
 \item{measured}{The measured quantification.}
 \item{reference}{The reference quantification.}
}
}
\description{
Note that both arguments must be phyloseq objects from the
phyloseq package.
}
\examples{
 NULL

}
