% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dada2.R
\name{with_classification}
\alias{with_classification}
\title{Quantify classification rates for each taxonomic rank.}
\usage{
with_classification(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A data table with 3 columns (rank, asvs, reads) specifying the
 fraction of ASVs or reads that were classified successfully on that
 taxa rank.
}
\description{
Quantify classification rates for each taxonomic rank.
}
