% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore.R
\name{align_nanopore}
\alias{align_nanopore}
\title{Align nanopore reads to a 16S reference database.}
\usage{
align_nanopore(read_files, ref, alignments_folder = "./alignments",
  log_file = "minimap2.log", threads = 4)
}
\arguments{
\item{read_files}{Paths to fastq files (can be gzipped).}

\item{ref}{Path to the reference fasta (can be gzipped).}

\item{alignments_folder}{Where to store the alignments.}

\item{log_file}{The log file used to store status messages.}

\item{threads}{Number of threds used for alignment.}
}
\value{
A data.table with sequence names, counts and sample name.
}
\description{
Align nanopore reads to a 16S reference database.
}
