% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributize.R
\name{attributize}
\alias{attributize}
\alias{attributize.default}
\alias{attributize.data.frame}
\alias{attributize.data.frame}
\alias{attributize.default}
\title{Map a vector to several plotting attributes}
\usage{
attributize(x, ...)

\method{attributize}{data.frame}(x, vec, vcol = ncol(x), ...)

\method{attributize}{default}(x, ..., na.last = FALSE, vcol = ncol(d))
}
\arguments{
\item{x, vec}{input vector}

\item{\dots}{other arguments passed to/from other methods}

\item{vcol}{numeric vector, column id with input vector values}

\item{na.last}{logical, passed to \code{sort}}
}
\value{
data frame
}
\description{
Given a vector with likely high number of unique values create a list of
vectors which combinations cover the unique values of the input vector.
}
