% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun.R
\name{count_hits}
\alias{count_hits}
\title{Count alignment hits to a reference database.}
\usage{
count_hits(alignment_files, tags = character(0), threads = 1)
}
\arguments{
\item{alignment_files}{Paths to BAM files.}

\item{tags}{Additional tags to read in the BAM file.}

\item{threads}{Number of parallel processes.}
}
\value{
A data.table with sequence names, counts and sample name.
}
\description{
Count alignment hits to a reference database.
}
