% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw.R
\name{cw}
\alias{cw}
\title{Reset the number of text columns of R console.}
\usage{
cw(w = Sys.getenv("COLUMNS"))
}
\arguments{
\item{w}{numeric, new width}
}
\value{
List with one component called \code{width} containing the old
setting.
}
\description{
Reset the number of text columns used by R console. By default reset to the
current width of the terminal using the \code{COLUMNS} environment variable.
}
\seealso{
\code{\link{options}}
}
