% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{demultiplex}
\alias{demultiplex}
\title{Splits FASTQ files into individual samples.}
\usage{
demultiplex(object, ...)
}
\arguments{
\item{object}{An experiment data table as returned by
\code{\link{find_read_files}} or a worflow object.}

\item{...}{A configuration as returned by
\code{\link{config_demultiplex}}.}
}
\value{
A list containing the split files and matching statistics.
}
\description{
This function will take each of the sequences barcodes and tries to find
the corresponding barcode from the reference set. After that it writes new
fastq files containing only the reads for a single barcode.
}
\examples{
 NULL

}
