% Generated by roxygen2 (4.0.2): do not edit by hand
\name{devlib}
\alias{devlib}
\title{Load package from developer's library}
\usage{
devlib(..., libloc = getOption("devlib"))
}
\arguments{
\item{libloc}{Path to developer's library}

\item{\dots}{other arguments passed to \code{\link{library}}}
}
\value{
Nothing of interest, but the package is loaded.
}
\description{
Load package from developer's library.
}
\seealso{
\code{\link{library}}, \code{\link{.libPaths}}
}

