% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_download.R
\name{download_files}
\alias{download_files}
\title{Download a list of files.}
\usage{
download_files(files, threads = 1)
}
\arguments{
\item{files}{A data frame-like object with three columns: url, target,
description specifying the source file, target location (including
the file name) and description.}
}
\value{
The list of files with indicated download success.
}
\description{
Download a list of files.
}
