% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ref.R
\name{filter_reference}
\alias{filter_reference}
\title{Filter a set of reference sequences from the data set.seed}
\usage{
filter_reference(reads, out, reference, alignments = NA)
}
\arguments{
\item{reads}{A data frame or data table containing the read files. Can be
generated with \link{\code{find_illumina}} for instance.}

\item{out}{The folder where to store filtered reads. Should be empty as
files **will be overwritten**.}

\item{reference}{Fasta file (can be gzipped) containing the reference
DNA sequences.}

\item{alignments}{Optional folder in which to store the alignments.}
}
\description{
This will also return a data.table containing the counts of reference
sequences for each sample.
}
