\name{mbtools-package}
\alias{mbtools-package}
\alias{mbtools}
\docType{package}
\title{Collection of development tools for \R}
\description{
This package is a collection of utility function for working with \R
and developing for \R.
}
\author{
Michal Bojanowski
\email{michal2992@gmail.com}
\url{http://bojan.3e.pl/weblog}
}
\seealso{
Other packages by Michal Bojanowski: \pkg{rgraph6}, \pkg{bojan}
}
\section{Changes}{

  \bold{Version 0.0-1} (2008-10-20)
  \itemize{
    \item First version of the package! Moved some utility
    functions out of package \pkg{bojan}.

    \item Added enhanced versions of \code{\link{setwd}},
    and \code{\link{getwd}}.
  }
}
\keyword{package}
