\docType{package}
\name{mbtools-package}
\alias{mbtools}
\alias{mbtools-package}
\title{Collection of development tools for R}
\description{
  This package is a collection of utility function for
  working with and developing for .
}
\author{
  Michal Bojanowski\cr \email{michal2992@gmail.com}\cr
  \url{http://www.bojanorama.pl}
}
\keyword{package}

