% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_plot}
\alias{mock_plot}
\title{Creates a plot of measured taxa quantifications vs. reference quantification.}
\usage{
mock_plot(ps, ref)
}
\arguments{
\item{ps}{A phyloseq object describing the measurements}

\item{ref}{A phyloseq object describing the reference set. Can be obtained
from \code{\link{mockrobiota}}.}
}
\value{
A ggplot2 plot.
}
\description{
Note that both arguments must be phyloseq objects from the
phyloseq package.
}
\examples{
 NULL

}
