% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mockrobiota}
\alias{mockrobiota}
\title{Downloads a complete data set from mockrobiota.}
\usage{
mockrobiota(name, folder, quiet = !interactive())
}
\arguments{
\item{name}{Name of the mockrobiota data set.}

\item{folder}{Where to save the data.}

\item{quiet}{Whether to show download progress.}
}
\value{
A list with the following components
\describe{
\item{description}{A description of the data set.}
\item{forward}{Filepath to the forward reads.}
\item{reverse}{Filepath to the reverse reads.}
\item{index}{Filepath to the index file.}
\item{citation}{Reference for the data set.}
\item{fragment}{Which fragment(s) were sequenced.}
\item{equipment}{The used sequencing equipment.}
\item{samples}{A data frame mapping samples to barcodes.}
\item{tax_gg}{The reference taxonomy for green genes.}
\item{tax_silva}{The reference taxonomy for silva.}
}
}
\description{
Downloads a complete data set from mockrobiota.
}
\examples{
 NULL

}

