% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{plot_qualities}
\alias{plot_qualities}
\title{Plots the quality profile for an entire experiment.}
\usage{
plot_qualities(qp, min_score = 10)
}
\arguments{
\item{qp}{A quality profile as returned by
\code{\link{quality_profile}}.}

\item{min_score}{Smallest acceptable score. Defaults to 10 (10% per base
error). A reference line will be plotted but no data will be discarded.}
}
\value{
A ggplot2 plot mapping the read positions to mean quality for each
 sample. Will be facetted into fowrward and reverse if applicable.
}
\description{
Plots the quality profile for an entire experiment.
}
