% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore.R
\name{read_bam}
\alias{read_bam}
\title{Read alignments from a BAM file.}
\usage{
read_bam(path, tags = character(0))
}
\arguments{
\item{path}{The file path to the BAM.}

\item{tags}{Additional tags to read from the BAM file.}
}
\value{
The alignments in a GAlignments object.
}
\description{
Read alignments from a BAM file.
}
