% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recoding variables (vectors)}
\usage{
recode(x, fromto, ...)
}
\arguments{
\item{x}{atomic vector, variable to be recoded}

\item{fromto}{two-column matrix or data frame, or a list}

\item{...}{Only used in the default method and expects further elements of
the recoding rule set}
}
\value{
Vector of the same mode as second column of \code{fromto} and length
as \code{x} with the values recoded.
}
\description{
This function has been moved to package \pkg{recode} at \url{https://github.com/mbojan/recode}.
}
\examples{
# Input vector
x <- c(1,2,3,4,3,2,1)

# Recoding with rules as a matrix
rmat <- matrix( c(1, 10, 4, 40), 2, 2, byrow=TRUE)
recode(x, rmat)

# Recoding with a data frame
d <- data.frame(from = c(2, 3),
                to = c("two", "three") )
recode(x, d)


# Recoding with rule sets
r <- recode(x, c(1,2), "one or two",
       c(3,4), "three or four",
       5, "five")
data.frame(x, r)
}
\seealso{
\code{\link{match}}
}
