% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ref.R
\name{remove_reference}
\alias{remove_reference}
\title{Removes sequence that map to a given reference.}
\usage{
remove_reference(reads, out, reference, alignments = NA, threads = 3)
}
\arguments{
\item{reads}{A character vector containing the read files in fastq format.
Can be generated using \code{\link{find_illumina}}.}

\item{out}{A folder to which to save the filtered fastq files.}

\item{reference}{Path to a fasta file (can be gzipped) that contains the
sequences to filter. Can be a genome or transcripts.}

\item{alignments}{Whether to keep the alignment. If not NA should be a
string indicating the path to the output bam file.}

\item{threads}{How many threads to use for mapping.}
}
\value{
A numeric vector with two entries. The number of sequences after
 filtering (non-mapped), and the number of removed sequences (mapped).
}
\description{
This can be used for a variety of applications the most common ones are:
\itemize{
  \item removing sequences from the host
  \item removing ribosomal sequences
  \item removing contaminants
}
This function uses minimap2 to align and identify hits and does not require
a prebuilt index.
}
\examples{
 NULL

}
