% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.keywords.R
\name{show.keywords}
\alias{show.keywords}
\title{Show the Keywords list}
\usage{
show.keywords(file = NULL)
}
\arguments{
\item{file}{path to file with keywords, defaults to
\file{RHOME/doc/KEYWORDS}}
}
\value{
Nothing, i.e. \code{NULL}.
}
\description{
This function will show the Keywords list needed for writing R Documentation
files in Rd format.
}
\details{
The function list the \code{file}. By default it is taken to be a
\file{KEYWORDS} file in \file{RHOME/doc/KEYWORDS}. The value of the RHOME
variable is taken by \code{Sys.getenv} function.
}
\examples{

show.keywords()

}
\seealso{
\code{\link{file.show}}, \code{\link{Sys.getenv}}
}
\keyword{file}
\keyword{utilities}
