% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb.R
\name{silva_build_taxonomy}
\alias{silva_build_taxonomy}
\title{Build the lineages for the entire SILVA database.}
\usage{
silva_build_taxonomy(taxmap, taxonomy)
}
\arguments{
\item{taxmap}{The SILVA taxonomy map (`taxmap_slv_*.txt`).}

\item{taxonomy}{The SILVA taxonomy.}
}
\value{
The SILVA lineage mapping every taxon to its upper ranks up to the
 root.
}
\description{
Build the lineages for the entire SILVA database.
}
