% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{slimm}
\alias{slimm}
\title{Quantifies abundances  and coverage using SLIMM.}
\usage{
slimm(object, ...)
}
\arguments{
\item{object}{An alignment table or workflow artifact that includes
alignments.}

\item{...}{Either additional arguments specifying a config or a config
created with \code{\link{config_slimm}}.}
}
\value{
A SLIMM artifact containing the lineage calls for each sample and
 taxon and the coverage.
}
\description{
Quantifies abundances  and coverage using SLIMM.
}
\examples{
 NULL

}
