% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun.R
\name{slimm}
\alias{slimm}
\title{Quantifies abundances for the bacteria using SLIMM}
\usage{
slimm(alignments, slimm_db)
}
\arguments{
\item{alignments}{A data frame as output by \code{\link{align_bowtie2}}}

\item{slimm_db}{Path for the SLIMM data base.}
}
\value{
A data table counting reads and relative abundances for all found
 bacteria for several taxonomic ranks. It will contain the following columns:
 \itemize{
 \item{id}{id of the sample}
 \item{rank}{the name of the taxonomic rank, for instance "genus"}
 \item{name}{the value of the rank, for instance "Escherichia"}
 \item{reads}{the number of reads in that rank}
 \item{relative}{the relative abundance of the rank in [0,1]}
 \item{coverage}{the coverage of the rank}
 }
}
\description{
Quantifies abundances for the bacteria using SLIMM
}
\examples{
 NULL

}
