% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangling.R
\name{taxa_count}
\alias{taxa_count}
\title{Count reads across a taxonomy level.}
\usage{
taxa_count(ps, lev = "Genus")
}
\arguments{
\item{ps}{A phyloseq object.}

\item{lev}{The taxonomy level to count.}
}
\value{
A mbquant data table with the taxa counts in long format.
}
\description{
Count reads across a taxonomy level.
}
\examples{
 NULL

}
