% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{taxa_metrics}
\alias{taxa_metrics}
\title{Calculates what percentage of taxa was found in a reference set.}
\usage{
taxa_metrics(taxa1, taxa2)
}
\arguments{
\item{taxa1}{First taxonomy table.}

\item{taxa2}{Second taxonomy table.}
}
\value{
A bar plot denoting the percentage of correctly identified taxa.
}
\description{
Calculates what percentage of taxa was found in a reference set.
}
\examples{
 NULL

}

