% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{taxa_quants}
\alias{taxa_quants}
\title{Compares taxa quantification from a measurement to a reference ground truth.}
\usage{
taxa_quants(taxa1, taxa2, normalize = FALSE)
}
\arguments{
\item{taxa1}{Measured taxonomy quantities.}

\item{taxa2}{Reference taxonomy quantities.}

\item{normalize}{Whether to normalize taxa counts first.}
}
\value{
A data frame with the following columns:
 \describe{
 \item{level}{The taxonomy level for the entry.}
 \item{name}{The taxonomy.}
 \item{measured}{The measured quantification.}
 \item{measured}{The reference quantification.}
}
}
\description{
Compares taxa quantification from a measurement to a reference ground truth.
}
\examples{
 NULL

}

