% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockrobiota.R
\name{taxa_quants}
\alias{taxa_quants}
\title{Compares taxa quantification from a measurement to a reference ground truth.}
\usage{
taxa_quants(taxa1, taxa2)
}
\arguments{
\item{taxa1}{Measured taxonomy quantities.}

\item{taxa2}{Reference taxonomy quantities.}
}
\value{
A scatter plot having the measured quantities on the x-axis and
 true quantities on the y-axis.
}
\description{
Compares taxa quantification from a measurement to a reference ground truth.
}
\examples{
 NULL

}

